actualizando

+++ src/main/webapp/detalle_genero.jsp
@@ -15,31 +15,30 @@
 %>
 
 <div class="container py-5">
-  <div class="row justify-content-center">
-    <div class="col-md-7">
-      <div class="card shadow-lg rounded-2">
-        <div class="card-header bg-black text-white text-center">
-          <h3 class="mb-0">Detalle del Género</h3>
-        </div>
-        <div class="card-body p-4">
-          <p><strong>Nombre:</strong> <%= genero.getNombre() %></p>
-          <p><strong>Descripción:</strong> <%= genero.getDescripcion() %></p>
-
-        </div>
-        <div class="card-footer text-end">
-          <a href="generos.jsp" class="btn btn-secondary btn-sm">Volver</a>
-          <%
-            if (role.equals("admin")){
-          %>
-          <a href="edit_genero.jsp?genero_id=<%= genero.getId() %>" class="btn btn-warning btn-sm">Editar</a>
-          <a href="delete_genero?genero_id=<%= genero.getId() %>" class="btn btn-danger btn-sm">Eliminar</a>
-          <%
-            }
-          %>
-        </div>
-      </div>
+    <div class="row justify-content-center">
+        <div class="col-md-7">
+            <div class="card shadow-lg rounded-2">
+                <div class="card-header bg-black text-white text-center">
+                    <h3 class="mb-0">Detalle del Género</h3>
+                </div>
+                <div class="card-body p-4">
+                    <p><strong>Nombre:</strong> <%= genero.getNombre() %></p>
+                    <p><strong>Descripción:</strong> <%= genero.getDescripcion() %></p>
+                </div>
+                <div class="card-footer text-end">
+                    <a href="generos.jsp" class="btn btn-secondary btn-sm">Volver</a>
+                    <%
+                        if (role.equals("admin")){
+                    %>
+                        <a href="edit_genero.jsp?genero_id=<%= genero.getId() %>" class="btn btn-warning btn-sm">Editar</a>
+                        <a href="delete_genero?genero_id=<%= genero.getId() %>" class="btn btn-danger btn-sm" onclick="return confirmar()">Eliminar</a>
+                    <%
+                        }
+                    %>
+                </div>
+              </div>
+          </div>
     </div>
-  </div>
 </div>
 
 <%
@@ -49,5 +48,6 @@
 <%
   }
 %>
+<script src="./scripts/script_confirmacion.js"></script>
 
 <%@include file="includes/footer.jsp"%>
